fix(sysinfo): cover the Windows collector, and fix four bugs it was hiding
WindowsServerInfoCollector was the last 0%-covered collector -- 1003 uncovered
lines, cognitive complexity 192. Unlike the Linux pair it needed no new seam: it
touches no files and reaches the OS entirely through exec(...), which was already
overridable. Commands arrive as one string, so the tests key canned responses on a
distinctive fragment of each command, which doubles as documentation of the WMI
class being stood in for.

51 tests, and they earned their keep immediately: the first run was 44/48, and the
four failures were exactly the two bugs predicted from reading the code.

- getRunningProcesses iterated processDataList.subList(0, 4) and
  getRunningServices subList(0, 5) -- both hardcoded. The admin UI therefore showed
  at most four processes and five services, and when the query returned fewer rows
  than the cap, subList threw IndexOutOfBoundsException straight into the
  method-level catch, yielding an empty list instead.

  Those caps weren't arbitrary: each item cost its own exec, so lifting them
  naively would mean one PowerShell invocation per service on a box with a couple
  of hundred. Adopt the bulk-lookup pattern the file already uses for CPU
  (getProcessCpuUsage): one Get-Process query builds a pid -> working-set map for
  services, and process-owner lookup runs only for the retained processes, now
  capped at the busiest 50 -- the same limit the Linux collector applies.

- Performance-counter routing compared paths case-sensitively
  (path.contains("Available MBytes")), but Get-Counter reports the resolved path
  lowercased. No branch ever matched, so availableMB, networkBytesIn/Out and the
  disk read/write counters were permanently zero.

- All three WMIC CSV fallbacks read columns by position, in the order the query
  listed them -- but wmic /format:csv emits Node first and then alphabetically.
  parseWmicCpuInfo put CurrentClockSpeed in `name` and then died on
  parseInt("AuthenticAMD"); parseWmicDiskInfo took DriveType as the size and
  FileSystem as the free space, so every row threw and the disk fallback returned
  nothing at all. Replaced with a header-name-driven parseWmicCsv, which is right
  under either column order rather than betting on one.

- Logical disks skipped CD-ROM and RAM disks only after adding their size to the
  running totals, so DiskInfo.totalSpace disagreed with the rows it summarised. The
  skip now precedes accumulation in both the CIM and WMIC paths, and a test pins
  totalSpace == sum(drives).

Also makes two implicit long-to-double widenings explicit (java:S2184 at
LinuxServerInfoCollector:812 and calculateDiskIOLoad). Neither was a real defect --
an epoch-millis delta guarded to under 60s, and a sum that would need ~2^63 to
overflow -- so this is a cast, not a restructure.

Verified: 719/719 backend (+51).

+++ db-world-backend/src/main/java/com/db/dbworld/app/system/info/collector/linux/LinuxServerInfoCollector.java
@@ -809,7 +809,7 @@ public class LinuxServerInfoCollector extends ServerInfoCollector {
         long prevTime = prevNetTimestamp;
         prevNetTimestamp = nowMs;
         if (prevTime > 0 && nowMs - prevTime < 60_000L && !prevNet.isEmpty()) {
-            double deltaMs = nowMs - prevTime;
+            double deltaMs = (double) nowMs - prevTime;
             for (Map.Entry<String, long[]> e : currentNet.entrySet()) {
                 long[] curr = e.getValue();
                 long[] prev = prevNet.getOrDefault(e.getKey(), new long[16]);
